+++ packages/dto/src/golf-genius/integration-log.dto.ts
@@ -11,6 +11,7 @@ export type IntegrationActionName =
 	| "Import Scores"
 	| "Import Points"
 	| "Import Results"
+	| "Import Team Results"
 	| "Import Skins"
 	| "Import Proxies"
 	| "Close Event"